+++ internal/service/error.go
@@ -29,6 +29,14 @@ const (
 )
 
 const (
+	SelectAuthErrMsg     = "could not select auth"
+	UpdateAuthErrMsg     = "could not update auth"
+	AuthNotFoundErrMsg   = "auth not found"
+	AuthAlreadyExistsMsg = "auth with the given external ID already exists"
+	AuthInvalidStatusMsg = "invalid auth status"
+)
+
+const (
 	MissingLabelKeysMsg = "missing label keys"
 	MissingLabelsMsg    = "missing labels"
 	EmptyLabelKeysMsg   = "label keys cannot be empty"
@@ -41,6 +49,7 @@ var (
 	ErrTenantIDFormat                   = status.Error(codes.InvalidArgument, "tenant ID is not valid")
 	ErrTenantNotFound                   = status.Error(codes.NotFound, TenantNotFoundMsg)
 	ErrTenantUnavailable                = status.Error(codes.FailedPrecondition, TenantUnavailableErrMsg)
+	ErrTenantEncoding                   = status.Error(codes.Internal, "failed to encode tenant data")
 	ErrTenantStatusTransitionNotAllowed = errors.New(TenantStatusTransitionNotAllowedMsg)
 	ErrInvalidTenantStatus              = errors.New(InvalidTenantStatusMsg)
 )
@@ -58,6 +67,14 @@ var (
 	ErrSystemUnavailable         = status.Error(codes.FailedPrecondition, SystemUnavailableErrMsg)
 	ErrNoSystemIdentifiers       = status.Error(codes.InvalidArgument, "no system identifiers provided")
 	ErrSystemListNotAllowed      = status.Error(codes.InvalidArgument, "need either externalID and region or tenantID to list systems")
+)
+
+var (
+	ErrAuthSelect        = status.Error(codes.Internal, SelectAuthErrMsg)
+	ErrAuthUpdate        = status.Error(codes.Internal, UpdateAuthErrMsg)
+	ErrAuthNotFound      = status.Error(codes.NotFound, AuthNotFoundErrMsg)
+	ErrAuthAlreadyExists = status.Error(codes.AlreadyExists, AuthAlreadyExistsMsg)
+	ErrAuthInvalidStatus = status.Error(codes.FailedPrecondition, AuthInvalidStatusMsg)
 )
 
 var (